Schedule (#9)
* schedule: create basic schedule message command without implementation

* schedule: create basic schedule renaming channel command without implementation

* schedule: enable strict null checks

* schedule: create service for fetching gifs. Add additional validation for scheduling message

* scheduled: create migration for scheduled messages

* schedule: log schedule message in channel

* scheduled: create migration for scheduled rename

* schedule: add renaming channel in db

* schedule: add pagination on scheduled messages

* schedule: add pagination on scheduled rename channels

* schedule: add cronjob for scheduling messages

* schedule: add cronjob for scheduling rename channels

+++ src/discord/modules/pagination/shadow-ban/shadow-ban-pagination.service.ts
@@ -39,8 +39,12 @@ export class ShadowBanPaginationService extends BasePaginationHandler<ShadowBanE
   public async showList(
     interaction: ChatInputCommandInteraction<CacheType>,
   ): Promise<void> {
+    const guildId = interaction.guildId;
+
+    if (!guildId) return;
+
     this.presets = await this.shadowBanRepository.find({
-      where: { guild: { guildId: interaction.guildId } },
+      where: { guild: { guildId: guildId } },
       relations: ['guild'],
     });
     this.totalPages = this.getTotalPages(this.presets, this.itemsPerPage);
@@ -181,16 +185,17 @@ export class ShadowBanPaginationService extends BasePaginationHandler<ShadowBanE
     embed: EmbedBuilder,
     btnInteraction: ButtonInteraction<CacheType>,
   ) {
+    const guildId = interaction.guildId;
+
+    if (!guildId) return;
+
     try {
       const pageDeletedInfo = this.getPageInfo(this.currentPage);
       await this.shadowBanRepository.delete({
         banId: pageDeletedInfo.pageData.banId,
       });
 
-      await this.clearShadowBanCache(
-        pageDeletedInfo.pageData.userIds,
-        interaction.guildId,
-      );
+      await this.clearShadowBanCache(pageDeletedInfo.pageData.userIds, guildId);
 
       await this.actionLoggerService.shadowBanRemove({
         name: pageDeletedInfo.pageData.name,
@@ -215,7 +220,7 @@ export class ShadowBanPaginationService extends BasePaginationHandler<ShadowBanE
       }
 
       this.presets = await this.shadowBanRepository.find({
-        where: { guild: { guildId: interaction.guildId } },
+        where: { guild: { guildId: guildId } },
         relations: ['guild'],
       });
       this.totalPages = this.getTotalPages(this.presets, this.itemsPerPage);
@@ -233,7 +238,7 @@ export class ShadowBanPaginationService extends BasePaginationHandler<ShadowBanE
       await this.updateEmbedOnBtnClick(embed, btnInteraction);
       return Promise.resolve();
     } catch (e) {
-      this.logger.error(`Shadow ban delete ${interaction.guildId}: ${e}`);
+      this.logger.error(`Shadow ban delete ${guildId}: ${e}`);
     }
   }
